okx and bithumb utils refactor
