+++ src/base/utils/utils.js
@@ -37,11 +37,7 @@ const timeIntervalToSeconds = (interval) => {
     return seconds;
 };
 
-const createHmacSignatureDefault = (
-    { ts, method, path, body = '' },
-    key,
-    encoding,
-) => {
+const createHmacSignatureDefault = ({ ts, method, path, body = '' }, key, encoding) => {
     if (body) body = JSON.stringify(body);
 
     const signaturePayload = [ts, method, path, body].join('');
@@ -49,8 +45,28 @@ const createHmacSignatureDefault = (
     return createHmac('sha256', key).update(signaturePayload).digest(encoding);
 };
 
+const createHmacSignatureBithumb = (sigParams, privateKey, encoding) => {
+    const charsToDel = ['{', '}', '"'];
+
+    let signaturePayload = JSON.stringify(sigParams)
+        .slice(1, -1)
+        .split(',')
+        .join('&')
+        .split(':')
+        .join('=');
+
+    charsToDel.forEach((item) => {
+        signaturePayload = signaturePayload.split(item).join('');
+    });
+
+    return createHmac('sha256', privateKey)
+        .update(signaturePayload)
+        .digest(encoding);
+};
+
 module.exports = {
     linkOriginalPayload,
     timeIntervalToSeconds,
     createHmacSignatureDefault,
+    createHmacSignatureBithumb,
 };
