+++ src/exchanges/bithumb/streams/marketDataStream.js
@@ -1,5 +1,5 @@
 const debug = require('../../../base/etc/debug');
-const utils = require('../utils');
+const utils = require('../../../base/utils/utils');
 
 const BithumbWebsocketBase = require('./websocketBase');
 const CandleStream = require('./additional/candleStream');
