+++ src/exchanges/bithumb/wallets/spot.js
@@ -1,11 +1,10 @@
 const BithumbBase = require('../base');
 const mergeObjects = require('deepmerge');
 
-const utils = require('../utils');
+const utils = require('../../../base/utils/utils');
 
 const AccountDataStream = require('../streams/accountDataStream');
 const MarketDataStream = require('../streams/marketDataStream');
-const { transformZenfuseOrder } = require('../utils');
 const { timeIntervals } = require('../metadata');
 
 /**
@@ -43,7 +42,7 @@ class BithumbSpot extends BithumbBase {
         });
         // TODO: Cache update here
 
-        const tickers = utils.extractSpotTickers(markets.data);
+        const tickers = markets.data.map((ticker) => ticker.s);
 
         utils.linkOriginalPayload(tickers, markets);
 
@@ -51,7 +50,7 @@ class BithumbSpot extends BithumbBase {
     }
 
     /**
-     * @returns {string[]} Array of ticker pairs on FTX
+     * @returns {string[]} Array of ticker pairs on Bithumb
      */
     async fetchMarkets() {
         const response = await this.publicFetch('spot/ticker', {
@@ -133,7 +132,7 @@ class BithumbSpot extends BithumbBase {
     async createOrder(zOrder) {
         this.validateOrderParams(zOrder);
 
-        const bOrder = transformZenfuseOrder(zOrder);
+        const bOrder = this.transformZenfuseOrder(zOrder);
 
         if (zOrder.type === 'market' && zOrder.side === 'buy') {
             let orderTotal = null;
@@ -150,7 +149,7 @@ class BithumbSpot extends BithumbBase {
             json: bOrder,
         });
 
-        const zCreatedOrder = utils.transformBithumbOrder(
+        const zCreatedOrder = this.transformBithumbOrder(
             bCreatedOrder,
             zOrder,
         );
@@ -267,7 +266,7 @@ class BithumbSpot extends BithumbBase {
             },
         });
 
-        const zOrder = utils.transformBithumbOrder(responce, orderToFetch);
+        const zOrder = this.transformBithumbOrder(responce, orderToFetch);
 
         return zOrder;
     }
@@ -331,6 +330,101 @@ class BithumbSpot extends BithumbBase {
         utils.linkOriginalPayload(result, response);
 
         return result;
+    }
+
+    /**
+     * @typedef {import('../../../../base/schemas/orderParams').ZenfuseOrderParams} OrderParams
+     */
+
+    /**
+     * Zenfuse -> Bithumb
+     *
+     * @param {OrderParams} zOrder
+     * @returns {object} Order for bithumb api
+     */
+    transformZenfuseOrder(zOrder) {
+        const TRANSFORM_LIST = ['side', 'type', 'price', 'quantity', 'symbol'];
+
+        const bOrder = {
+            symbol: zOrder.symbol.replace('/', '-'),
+            type: zOrder.type,
+            side: zOrder.side,
+            quantity: zOrder.quantity.toString(),
+            timestamp: Date.now().toString(),
+        };
+
+        if (zOrder.price) {
+            bOrder.price = zOrder.price.toString();
+        }
+
+        if (zOrder.type === 'market') {
+            bOrder.price = '-1';
+        }
+
+        // Allow user extra keys
+        for (const [key, value] of Object.entries(zOrder)) {
+            if (!TRANSFORM_LIST.includes(key)) {
+                bOrder[key] = value;
+            }
+        }
+
+        return bOrder;
+    }
+
+    /**
+     * @typedef {import('../../../../base/schemas/openOrder').PlacedOrder} PlacedOrder
+     */
+
+    /**
+     * Bithumb -> Zenfuse
+     *
+     * @param {*} bOrder Order from Bithumb REST
+     * @param {object} zInitialOrder
+     * @returns {PlacedOrder} Zenfuse Order
+     */
+    transformBithumbOrder(bOrder, zInitialOrder = {}) {
+        /**
+         * @type {PlacedOrder}
+         */
+        const zOrder = {};
+
+        // TODO: Refactor this
+
+        zOrder.id = bOrder.data.orderId;
+        if (Object.entries(zInitialOrder).length === 0) {
+            //if order is not cached
+            if (bOrder.data.status === 'success') {
+                zOrder.status = 'close';
+            } else if (
+                bOrder.data.status === 'send' ||
+                bOrder.data.status === 'pending'
+            ) {
+                zOrder.status = 'open';
+            } else {
+                zOrder.status = 'canceled';
+            }
+            zOrder.symbol = bOrder.data.symbol.replace('-', '/');
+            zOrder.timestamp = bOrder.timestamp;
+            zOrder.type = bOrder.data.type;
+            zOrder.side = bOrder.data.side;
+            zOrder.price = bOrder.data.price
+                ? parseFloat(bOrder.data.price)
+                : undefined;
+            zOrder.quantity = parseFloat(bOrder.data.quantity);
+        } else {
+            zOrder.symbol = zInitialOrder.symbol;
+            zOrder.timestamp = zInitialOrder.timestamp
+                ? zInitialOrder.timestamp
+                : Date.now();
+            zOrder.type = zInitialOrder.type;
+            zOrder.side = zInitialOrder.side;
+            zOrder.quantity = zInitialOrder.quantity;
+            zOrder.price = zInitialOrder.price ? zInitialOrder.price : undefined;
+            zOrder.status = zInitialOrder.status ? zInitialOrder.status : 'open';
+        }
+        // zOrder.trades = bOrder.fills; // TODO: Fill commision counter
+
+        return zOrder;
     }
 
     getAccountDataStream() {
