+++ src/exchanges/okx/errors/api.error.js
@@ -1,5 +1,5 @@
 const ExchangeBaseException = require('../../../base/errors/exchange.error');
-const utils = require('../utils');
+const utils = require('../../../base/utils/utils');
 
 const codes = ExchangeBaseException.errorCodes;
