--- src/exchanges/okx/utils/functions/agregation.js
@@ -1,1 +1,0 @@
-module.exports = {};
