--- src/exchanges/okx/utils/index.js
@@ -1,7 +1,0 @@
-module.exports = {
-    ...require('../../../base/utils/utils'),
-    ...require('./functions/agregation'),
-    ...require('./functions/transformation'),
-    ...require('./functions/validatation'),
-    ...require('./functions/other'),
-};
